iqoption_startup_project/requirements.txt
git+https://github.com/Lu-Yi-Hsun/iqoptionapi.git@master#egg=iqoptionapi
python-dotenv>=0.21.1
pandas>=1.5.3
numpy>=1.24.3
urllib3>=1.26.16,<2.0.0
requests>=2.28.2
websocket-client>=1.5.1
python-dateutil>=2.8.2
pytz>=2023.3
